refactor: getWishList endpoint /clothes->/wish로 변경
