+++ src/controller/clothes/router.ts
@@ -4,7 +4,6 @@ import {
   getClothes,
   modifyClothes,
   deleteClothes,
-  getWishList,
 } from './controller';
 
 const clothesRouter = Router();
@@ -13,6 +12,5 @@ clothesRouter.post('/', createClothes);
 clothesRouter.put('/:clothesId', modifyClothes);
 clothesRouter.get('/:clothesId', getClothes);
 clothesRouter.delete('/:clothesId', deleteClothes);
-clothesRouter.get('/', getWishList);
 
 export default clothesRouter;
