+++ src/controller/wish/controller.ts
@@ -4,6 +4,7 @@ import WishService from '../../service/wish.service';
 import CreateWishReq from '../../type/wish/createWish.req';
 import DefaultRes from '../../type/default.res';
 import LoginUser from '../../type/user/loginUser';
+import GetWishListRes from '../../type/wish/getWishList.res';
 
 export const createWish: RequestHandler = async (req, res, next) => {
   try {
@@ -51,6 +52,21 @@ export const deleteWish: RequestHandler = async (req, res, next) => {
 
     const message: DefaultRes = { message: '찜이 삭제되었습니다.' };
     res.json(message);
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getWishList: RequestHandler = async (req, res, next) => {
+  try {
+    const user = req.user as LoginUser;
+
+    if (!user) {
+      throw new UnauthorizedError('로그인이 필요한 기능입니다.');
+    }
+
+    const wishList: GetWishListRes = await WishService.getWishList(user.id);
+    res.json(wishList);
   } catch (error) {
     next(error);
   }
